+++ src/MovieBot.Core/ManifestJson.cs
@@ -16,6 +16,8 @@ namespace TheKrystalShip.MovieBot.Core;
 [JsonSerializable(typeof(Manifest))]
 [JsonSerializable(typeof(SidecarSubtitle))]
 [JsonSerializable(typeof(SubtitlePin))]
+[JsonSerializable(typeof(Chapter))]
+[JsonSerializable(typeof(ThumbnailStrip))]
 [JsonSerializable(typeof(SessionState))]
 [JsonSerializable(typeof(SessionStatePush))]
 [JsonSerializable(typeof(SeekClamped))]